+++ packages/qb/pqb/src/queryMethods/queryMethods.ts
@@ -69,9 +69,7 @@ type WindowResult<T extends Query, W extends WindowArg<T>> = SetQueryWindows<
 export type OrderArg<T extends Query> =
   | keyof T['selectable']
   | {
-      [K in Selectable<T>]?:
-        | SortDir
-        | { dir: SortDir; nulls: 'FIRST' | 'LAST' };
+      [K in Selectable<T>]?: SortDir;
     }
   | RawExpression;